Add files via upload
In this python plotting project, I was given 4 options and I had the freedom to choose the option I feel more confident in. Therefore, I chose the easiest option, which asked me to create a plot where I had to create a reference value and if the bars' value was smaller, they would be blue, if they were bigger, they would be red and, finally, if the bars contained the reference value, they would be white.